style: split ActiveSuggestionModel into its own file, tighten comments
Align #7665 changes with code-style.instructions.md: one public class
per file, and comment blocks trimmed to the minimal-comment convention
(contract rationale kept, restated behavior cut). No logic changes.
